Add sync line to Spectrum plot as well.
